Organized file structure for Database CRUD functions
Created a new directory that contains the code logic for the CRUD
functions for the FirebaseDB and created a new file to store the
location class. Organized the code so database functions are in
separate files and empties MainActivity.java for UI code logic.
